+++ fiddles/svg/fiddle-0027-ImageCloud/src/metadata.js
@@ -1,3 +1,5 @@
 app.metadata = app.metadata || {
-    gitHubUrl: 'https://github.com/bradyhouse/house/tree/master/fiddles/svg/fiddle-0027-ImageCloud'
+    consoleTag: 'H O U S E ~ f i d d l e s',
+    gitHubUrl: 'https://github.com/bradyhouse/house/tree/master/fiddles/svg/fiddle-0027-ImageCloud',
+    dataUrl: 'https://s3-us-west-2.amazonaws.com/s.cdpn.io/297733/photo-album.json'
   };